requirements.txt
python-telegram-bot
python-dotenv
telethon
psutil
requests
wikipedia
pyrogram
pillow
spamwatch
telegraph
pymongo
redis
Flask
# Add any other dependencies your module1 need below
